requirements.txt
# Core Python Libraries
numpy==1.24.3
pandas==2.0.3
scipy==1.11.4
pyyaml

# RAG Pipeline - ML & Embeddings
torch>=2.0.0,<2.3.0
sentence-transformers==2.7.0
transformers==4.38.0
tokenizers>=0.15.0

# RAG Pipeline - Vector Search
faiss-cpu==1.8.0

# RAG Pipeline - Web Scraping
beautifulsoup4==4.12.2
playwright==1.40.0
aiohttp==3.9.1
newspaper3k
lxml
trafilatura
markdownify

# RAG Pipeline - NLP and Text Processing
tiktoken==0.5.2
nltk==3.8.1

# Vision Pipeline - Image Processing
Pillow
kagglehub
kaggle

# Vision Pipeline - Data Generation & Validation
Faker
pyarrow

# Shared - Data Validation & Metadata
great-expectations>=0.18.0
mlflow

# Shared - Fairness and Bias Detection
scikit-learn
fairlearn
tensorflow
# tensorflow-model-analysis  # Optional

# Cloud & Storage
google-cloud-storage>=2.10.0
dvc[gs]

# Testing Framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-timeout>=2.1.0
pytest-xdist>=3.3.1
pytest-env>=0.8.2
pytest-html>=3.2.0
pytest-json-report>=1.5.0
pytest-benchmark>=4.0.0
pytest-doctestplus>=1.0.0

# Mocking and Test Fixtures
mock>=5.1.0
factory-boy>=3.3.0
hypothesis>=6.82.0

# Code Quality and Coverage
coverage>=7.3.0

# Code Quality & Linting
pylint>=2.17.5
flake8>=6.1.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.1